requirements.txt
bottle
harmony-connect-client
python-identitykeys